requirements.txt
tapioca-wrapper<2
requests-oauthlib
